+++ doc/changelog.md
@@ -1,3 +1,9 @@
+v2.5.7
+======
+
+- Fixes problems handling the radius and zoom preferences
+- Add license metadata
+
 v2.5.5
 ======